requirements.txt
numpy
tensorflow
keras
pillow
cython